argumentクラス、file_listクラス、formatモジュールに分割
