+++ 05.ls/class
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require 'pathname'
+require 'byebug'
+
+require './argument'
+
+class ShortFormat
+
